requirements.txt
numpy>=1.21.0
Pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
openai>=0.27.0
opencv-python>=4.5.0
scipy>=1.7.0
matplotlib>=3.4.0
tqdm>=4.62.0

# Optional dependencies for REFLECT integration
zarr>=2.10.0
imagecodecs>=2021.0.0
open3d>=0.13.0
scikit-learn>=0.24.0

# Optional dependencies for AI2THOR simulation
# ai2thor>=4.0.0


